+++ lib/src/t_server/t_server_http_listener.dart
@@ -1,9 +1,27 @@
 import 'dart:io';
 
+///
+/// server http listener
+///
 class TServerHttpListener {
+  ///
+  /// url path
+  ///
   String url;
+
+  ///
+  /// GET,POST,DELETE,PUT
+  ///
   String method;
+
+  ///
+  /// listener callback
+  ///
   void Function(HttpRequest req) onRequest;
+
+  ///
+  /// constructor
+  ///
   TServerHttpListener({
     required this.url,
     required this.method,