+++ lib/src/t_server/t_server_socket_listener.dart
@@ -1,7 +1,19 @@
 import 'dart:io';
 
+///
+/// server socket listener
+///
 class TServerSocketListener {
+  ///
+  /// socket url path
+  ///
   String url;
+  ///
+  /// socket listener
+  ///
   void Function(HttpRequest req, WebSocket socket) onListen;
+  ///
+  /// constructor
+  ///
   TServerSocketListener({required this.url, required this.onListen});
 }